adding remove handler
